pyproject.toml
[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
requires-python = ">=3.8"
dependencies = [
    "croniter~=0.3.34",
    "django~=3.1.0",
    "django-allauth~=0.44.0",
    "django-anymail~=8.1.0",
    "django-configurations[cache,database,email]~=2.2.0",
    "django-debug-toolbar~=3.2.0",
    "django-extensions~=3.1.0",
    "django-health-check~=3.16.2",
    "django-mail-panel~=3.1.0",
    "django-settings-export~=1.2.1",
    "django-q~=1.3.3",
    "django-redis~=4.12.1",
    "django-storages~=1.11.0",
    "django-widget-tweaks~=1.4.8",
    "factory_boy~=3.1.0",
    "gunicorn~=20.0.4",
    "pillow~=8.0.1",
    "psycopg2~=2.8.6",
    "watchdog[watchmedo]~=1.0.2",
    "werkzeug[watchdog]~=1.0.1",
]
dev-dependencies = [
    "autopep8~=1.5.0",
    "black==20.8b1",
    "flake8~=3.8.0",
    "flake8-assertive~=1.3.0",
    "flake8-bugbear~=20.11.1",
    "flake8-builtins~=1.5.2",
    "flake8-commas~=2.0.0",
    "flake8-comprehensions~=3.3.1",
    "flake8-datetimez~=20.10.0",
    "flake8-pytest-style~=1.3.0",
    "flake8-quotes~=3.2.0",
    "flake8-isort~=4.0.0",
    "isort~=5.6.4",
    "pylint~=2.6.0",
    "pylint-django~=2.3.0",
    "pylint-runner~=0.6.0",
    "pytest~=6.2.1",
    "pytest-cov~=2.10.0",
    "pytest-django~=4.1.0",
    "pytest-only~=1.2.2",
    "pytest-watch~=4.2.0",
    "pyupgrade~=2.7.0",
]

[tool.autopep8]
max_line_length = 80

[tool.black]
line-length = 79
target-version = ['py38']

[tool.isort]
known_django = ["django"]
src_paths = ["src"]
profile = "django"
force_single_line = true
length_sort = true
sections = ["STDLIB", "DJANGO", "THIRDPARTY" , "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
addopts = "-ra --cov=backend --cov-config=.coveragerc --cov-report term-missing"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
DJANGO_SETTINGS_MODULE = "core.settings"
DJANGO_CONFIGURATION = "Testing"

[tool.pdm]
[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true

[tool.pdm.scripts]
"lint:black" = "black --check backend"
"lint:flake8" = "flake8 backend --count"
"lint:pylint" = "sh -c \"DJANGO_SETTINGS_MODULE=core.settings DJANGO_CONFIGURATION=Base pdm run pylint backend/**/*.py\""
"lint:fix:autopep8" = "autopep8 -ria backend"
"lint:fix:black" = "black backend"
"lint:fix:isort" = "isort backend"
"lint:fix:pyupgrade" = "sh -c \"pdm run pyupgrade --py38-plus backend/**/*.py\""
"test" = "py.test backend --ds=core.settings --dc=Testing"
"test:watch" = "ptw backend --ds=core.settings --dc=Testing"

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
